soln

+++ wallis.c
@@ -27,7 +27,7 @@ int main(void) {
 float  wallis_pi(int n){
   float y,x=1.00; 
   for(int a=1 ; a<=n ; a++){
-    y = (4*a*a) / (4*a*a -1);
+    y =(float) (4*a*a) / (4*a*a -1);
     x=x*y;
   }
   x=2*x;
